refactor: update type module to cross compile Go plugins
