+++ cmd/tracee-rules/output_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/aquasecurity/tracee/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
