+++ signatures/golang/export.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/aquasecurity/tracee/tracee-rules/types"
+import "github.com/aquasecurity/tracee/types"
 
 // ExportedSignatures fulfills the goplugins contract required by the rule-engine
 // this is a list of signatures that this plugin exports
