+++ signatures/golang/stdio_over_socket_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/aquasecurity/tracee/pkg/external"
 	"github.com/aquasecurity/tracee/signatures/signaturestest"
-	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/aquasecurity/tracee/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
