+++ tracee-rules/benchmark/signature/golang/anti_debugging.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/aquasecurity/tracee/pkg/external"
 	"github.com/aquasecurity/tracee/signatures/helpers"
-	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/aquasecurity/tracee/types"
 )
 
 type antiDebugging struct {
