+++ tracee-rules/benchmark/signature/golang/code_injection.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/aquasecurity/tracee/pkg/external"
 	"github.com/aquasecurity/tracee/signatures/helpers"
-	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/aquasecurity/tracee/types"
 )
 
 type codeInjection struct {
