+++ tracee-rules/engine/engine_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/aquasecurity/tracee/pkg/external"
-	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/aquasecurity/tracee/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
